@@ -190,7 +190,7 @@ namespace Sanakan.Modules
 
             using (var db = new Database.UserContext(Config))
             {
-                var botuser = await db.Users.Where(x => x.Id == usr.Id).AsNoTracking().FirstOrDefaultAsync();
+                var botuser = await db.Users.AsQueryable().Where(x => x.Id == usr.Id).AsNoTracking().FirstOrDefaultAsync();
                 if (botuser == null)
                 {
                     await ReplyAsync("", embed: "Ta osoba nie ma profilu bota.".ToEmbedMessage(EMType.Error).Build());
